data: Move batch review selection freshness
Batch file selection translation needs a freshness check before it maps
review gutter IDs to batch selection IDs.

The check lived in action scope, whose remaining public API resolves
command scope and validates pathless line actions. Keeping batch-selection
freshness there made action scope carry a helper owned by a narrower
caller.

This commit addresses that by moving fresh batch-review selection lookup
into a batch-selection freshness module and routing batch selection through
it.

Subsequent architecture coverage will record the new file-review boundary.
