+++ src/git_stage_batch/data/file_review/batch_selection.py
@@ -18,7 +18,7 @@ from ...batch.submodule_pointer import (
 from ...core.line_selection import LineRanges
 from ...exceptions import exit_with_error
 from ...i18n import _
-from .action_scope import fresh_batch_review_selections_for_action
+from .batch_selection_freshness import fresh_batch_review_selections_for_action
 from .records import FileReviewAction
 from .selection_validation import validate_review_scoped_line_selection
